Move batch tests under inscribe
