+++ tests/wallet.rs
@@ -1,7 +1,6 @@
 use super::*;
 
 mod balance;
-mod batch_inscribe;
 mod cardinals;
 mod create;
 mod inscribe;
